Add Token counts fixes. Add decimal prop for CountUp. Organised and separated FarmBanner files
